requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
bcrypt==4.1.2
certifi==2025.8.3
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.110.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
loguru==0.7.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.3.4
pydantic_core==2.23.4
pytest==8.1.1
pytest-asyncio==0.23.5
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.27.1
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
